chore(): new tests + set up in the wild mani testing
