+++ libraries/manifest-validation/package.json
@@ -15,7 +15,10 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "typescript": "^4.6.3",
-    "rimraf": "^3.0.2"
+    "rimraf": "^3.0.2",
+    "typescript": "^4.6.3"
+  },
+  "dependencies": {
+    "node-fetch": "^3.3.0"
   }
 }
